updated to parsefloat not parseinst

+++ src/lambda/ai-request/index.mjs
@@ -3,8 +3,8 @@ import { InvokeCommand, LambdaClient } from "@aws-sdk/client-lambda";
 const lambdaClient = new LambdaClient({});
 
 export async function handler(event) {
-  const lat = Number.parseInt(event.queryStringParameters.lat);
-  const long = Number.parseInt(event.queryStringParameters.long);
+  const lat = Number.parseFloat(event.queryStringParameters.lat);
+  const long = Number.parseFloat(event.queryStringParameters.long);
   const s3Key = (Date.now() + Math.random()).toString();
 
   const invokeCmd = new InvokeCommand({
